+++ src/imio/smartweb/core/contents/sections/news/view.py
@@ -28,6 +28,7 @@ class NewsView(CarouselOrTableSectionView, HashableJsonSectionView):
             "portal_type=imio.news.NewsItem",
             "metadata_fields=container_uid",
             "metadata_fields=category_title",
+            "metadata_fields=local_category",
             "metadata_fields=topics",
             "metadata_fields=has_leadimage",
             "metadata_fields=modified",
@@ -60,7 +61,7 @@ class NewsView(CarouselOrTableSectionView, HashableJsonSectionView):
             modified_hash = hash_md5(item["modified"])
             category = ""
             if self.context.show_categories_or_topics == "category":
-                category = item.get("category_title", "")
+                category = item.get("local_category") or item.get("category_title", "")
             elif self.context.show_categories_or_topics == "topic":
                 topic = item.get("topics") and item["topics"][0] or None
                 category = translate_vocabulary_term(