+++ src/imio/smartweb/core/tests/test_vocabularies.py
@@ -265,7 +265,7 @@ class TestVocabularies(ImioSmartwebTestCase):
         json_agendas_raw_mock = get_json("resources/json_events_agendas_raw_mock.json")
         url = f"{config.EVENTS_URL}/imio-events-entity/@search?portal_type=imio.events.Agenda&sort_on=sortable_title&b_size=1000000&metadata_fields=UID"
         m.get(url, text=json.dumps(json_agendas_raw_mock))
-        url = f"{config.EVENTS_URL}/@search?selected_agendas=64f4cbee9a394a018a951f6d94452914&selected_agendas=96d3e3299dc74386943e12c4f4fd0b8a&portal_type=imio.events.Event&metadata_fields=category&metadata_fields=topics&metadata_fields=start&metadata_fields=end&metadata_fields=has_leadimage&metadata_fields=breadcrumb&metadata_fields=UID&event_dates.query=2021-11-15&event_dates.range=min&b_size=1000000"
+        url = f"{config.EVENTS_URL}/@search?selected_agendas=64f4cbee9a394a018a951f6d94452914&selected_agendas=96d3e3299dc74386943e12c4f4fd0b8a&portal_type=imio.events.Event&metadata_fields=category_title&metadata_fields=topics&metadata_fields=start&metadata_fields=end&metadata_fields=has_leadimage&metadata_fields=breadcrumb&metadata_fields=UID&event_dates.query=2021-11-15&event_dates.range=min&b_size=1000000"
         json_rest_events = get_json("resources/json_rest_events_with_breadcrumbs.json")
         m.get(url, text=json.dumps(json_rest_events))
         vocabulary = get_vocabulary("imio.smartweb.vocabulary.EventsFromEntity")
@@ -284,7 +284,7 @@ class TestVocabularies(ImioSmartwebTestCase):
         )
         url = f"{config.NEWS_URL}/imio-news-entity/@search?portal_type=imio.news.NewsFolder&sort_on=sortable_title&b_size=1000000&metadata_fields=UID"
         m.get(url, text=json.dumps(json_newsfolders_raw_mock))
-        url = f"{config.NEWS_URL}/@search?selected_news_folders=64f4cbee9a394a018a951f6d94452914&selected_news_folders=96d3e3299dc74386943e12c4f4fd0b8a&portal_type=imio.news.NewsItem&metadata_fields=category&metadata_fields=topics&metadata_fields=has_leadimage&metadata_fields=breadcrumb&metadata_fields=UID&b_size=1000000"
+        url = f"{config.NEWS_URL}/@search?selected_news_folders=64f4cbee9a394a018a951f6d94452914&selected_news_folders=96d3e3299dc74386943e12c4f4fd0b8a&portal_type=imio.news.NewsItem&metadata_fields=category_title&metadata_fields=topics&metadata_fields=has_leadimage&metadata_fields=breadcrumb&metadata_fields=UID&b_size=1000000"
         json_rest_news = get_json("resources/json_rest_news.json")
         m.get(url, text=json.dumps(json_rest_news))
         vocabulary = get_vocabulary("imio.smartweb.vocabulary.NewsItemsFromEntity")